TIKA-831 Fix the data type when comparing errors from the forked server, and add some more Forked unit tests (one disabled) - patch originally from Jerome Lacoste
git-svn-id: https://svn.apache.org/repos/asf/tika/trunk@1224728 13f79535-47bb-0310-9956-ffa450edef68

+++ tika-core/src/main/java/org/apache/tika/fork/ForkClient.java
@@ -172,7 +172,7 @@ class ForkClient {
                 ForkResource resource =
                     resources.get(input.readUnsignedByte());
                 resource.process(input, output);
-            } else if (type == ForkServer.ERROR) {
+            } else if ((byte) type == ForkServer.ERROR) {
                 try {
                     return (Throwable) ForkObjectInputStream.readObject(
                             input, loader);
